@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.1+2
 
 environment:
-  sdk: ">=2.16.2 <3.0.0"
+  sdk: '>=2.19.2 <3.0.0'
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -34,16 +34,32 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
-  firebase_core:
-  animated_splash_screen:
-  firebase_auth:
+  google_fonts:
   cloud_firestore:
+  firebase_core:
+  firebase_auth:
+  connectivity_plus:
+  bloc:
+  flutter_bloc:
+  bloc_concurrency:
+  equatable:
+  shared_preferences:
+  country_picker:
+  country_code_picker:
+  font_awesome_flutter:
+  fluttertoast: ^8.1.3
+  modal_bottom_sheet:
+  animated_splash_screen:
+  pinput:
+  country_list_picker:
   provider:
-  fluttertoast:
   carousel_slider:
   cached_network_image:
-  font_awesome_flutter:
+  intl:
   url_launcher:
+  image_picker:
+  firebase_storage:
+  flutter_typeahead:
 
 dev_dependencies:
   flutter_test:
