@@ -1,20 +1,28 @@
 class Assets {
-  static const String btc = "assets/images/btc.png";
-  static const String expoinLogo = 'assets/images/expoin logo.jpg';
-  static const String info1 = 'assets/images/computer science.jpg';
-  static const String info2 = 'assets/images/info2.jpg';
-  static const String ada = 'assets/images/ada.png';
-  static const String bnb = 'assets/images/bnb.png';
-  static const String eth = 'assets/images/eth.jpg';
-  static const String usdt = 'assets/images/usdt.png';
-  static const String xrp = 'assets/images/xrp.jpg';
-  static const String facebookLogo = 'assets/images/facebook logo.png';
-  static const String twitterLogo = 'assets/images/twitter-logo.png';
-  static const String youtubeLogo = 'assets/images/youtube logo.png';
-  static const String whatsappLogo = 'assets/images/whatsapp logo.png';
-  static const String telegramLogo = 'assets/images/telegram_logo.png';
-  static const String airtelMoney = 'assets/images/airtel money.png';
-  static const String orangeMoney = 'assets/images/orange money.png';
-  static const String mpesa = 'assets/images/m-pesa.png';
+  static const String btcImage = 'assets/images/btc.png';
+  static const String adaImage = 'assets/images/ada.png';
+  static const String bnbImage = 'assets/images/bnb.png';
+  static const String cosmosAtomImage = 'assets/images/cosmos-atom-logo.png';
+  static const String dogecoinImage = 'assets/images/dogecoin-doge-logo.png';
+  static const String ethImage = 'assets/images/eth.jpg';
+  static const String ethereumClassicImage = 'assets/images/ethereum-classic-etc-logo.png';
+  static const String iconIcxImage = 'assets/images/icon-icx-logo.png';
+  static const String komodoImage = 'assets/images/komodo-kmd-logo.png';
+  static const String solanaImage = 'assets/images/solana-sol-logo.png';
+  static const String tezosImage = 'assets/images/tezos-xtz-logo.png';
+  static const String tronTrxImage = 'assets/images/tron-trx-logo.png';
+  static const String usdtImage = 'assets/images/usdt.png';
+  static const String xrpImage = 'assets/images/xrp.jpg';
+  static const String zilligaImage = 'assets/images/zilliqa-zil-logo.png';
+  static const String bandProtocol = 'assets/images/band-protocol-band-logo.png';
+  static const String teraImage = 'assets/images/tera-tera-logo.png';
+  static const String tetherImage = 'assets/images/tether-usdt-logo.png';
+  static const String shibaInuImage = 'assets/images/shiba inu.png';
+  static const String airtelLogo = 'assets/images/airtel money.png';
+  static const String mpesaLogo = 'assets/images/m-pesa.png';
+  static const String orangeLogo = 'assets/images/orange money.png';
 
+  static const String mesPiecesLogo = 'assets/images/mes pieces logo.jpg';
+  static const String noDataImage = 'assets/images/no data.jpg';
+  static const String profileImage = 'assets/images/profile.jpeg';
 }
